+++ packages/frontend/src/components/common/Dashboard/DashboardHeaderV2.tsx
@@ -49,6 +49,7 @@ import { useProject } from '../../../hooks/useProject';
 import useApp from '../../../providers/App/useApp';
 import useTracking from '../../../providers/Tracking/useTracking';
 import { EventName } from '../../../types/Events';
+import { SwitchToClassicMenuItem } from '../../../features/dashboardTabsV2/DashboardUIToggle';
 import AddTileButton from '../../DashboardTiles/AddTileButton';
 import MantineIcon from '../MantineIcon';
 import PageHeader from '../Page/PageHeader';
@@ -580,6 +581,10 @@ const DashboardHeaderV2 = ({
                                         Export dashboard
                                     </Menu.Item>
                                 )}
+
+                                <Menu.Divider />
+
+                                <SwitchToClassicMenuItem />
 
                                 {userCanManageDashboard && (
                                     <>